Add files via upload
Due to poor dietary habits, inadequate sleep, and a lack of daily activity, the majority of individuals in the modern world are disease-prone. But for a doctor, making an accurate forecast based on symptoms is too challenging.
The hardest task is making an accurate diagnosis of a condition. Each year, there is significant data increase in the medical sciences. The accurate analysis of medical data, which has benefited from early  patient care, has increased as a result of the expansion in data in the  medical and healthcare fields. Using machine learning, created a method  to forecast diseases. These systems will output the potential disease that an individual may have based on their symptoms.